requirements.txt
customtkinter~=5.2.2
coloredlogs~=15.0.1
pyinstaller~=6.11.1
tkinterdnd2~=0.3.0
pillow~=11.1.0